MGSAPI/requirements.txt
alembic==1.3.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.1.0
pbr==5.4.4
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
stevedore==1.31.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.16.0
zipp==1.0.0
